+++ CacheSystem/CacheDataStructure.h
@@ -18,12 +18,12 @@ namespace CacheSystem
 		/**
 		map of vectors of the data
 		*/
-		std::unordered_map<uint64_t, std::vector<std::shared_ptr<CacheData> > > cacheData;
+		std::unordered_map<size_t, std::vector<std::shared_ptr<CacheData> > > cacheData;
 		
 		/**
 		iterator of the cacheData map
 		*/
-		std::unordered_map<uint64_t, std::vector<std::shared_ptr<CacheData> > >::iterator mapIterator;
+		std::unordered_map<size_t, std::vector<std::shared_ptr<CacheData> > >::iterator mapIterator;
 
 		/**
 		iterator for the list that is currently selected by the mapIterator
@@ -34,13 +34,13 @@ namespace CacheSystem
 		/**
 		finds and returns the CacheData object in which all the input parameters are equal to the corresponding input parameters passed as params
 		*/
-		template <class... Types> std::shared_ptr<CacheData> getCacheData(uint64_t hash, const std::vector<std::shared_ptr<ParameterInfo> > & paramsInfo,
+		template <class... Types> std::shared_ptr<CacheData> getCacheData(size_t hash, const std::vector<std::shared_ptr<ParameterInfo> > & paramsInfo,
 			void* dependencyObject, const Types &... params);
 
 		/**
 		adds new cache data, the CacheData object must be created by the new operator and it will be deleted during deleting this object
 		*/
-		void addCacheData(uint64_t hash, std::shared_ptr<CacheData> data);
+		void addCacheData(size_t hash, std::shared_ptr<CacheData> data);
 
 		/**
 		resets the data iterator
@@ -59,7 +59,7 @@ namespace CacheSystem
 	};
 
 	template <class... Types>
-	std::shared_ptr<CacheData> CacheDataStructure::getCacheData(uint64_t hash, const std::vector<std::shared_ptr<ParameterInfo> > & paramsInfo,
+	std::shared_ptr<CacheData> CacheDataStructure::getCacheData(size_t hash, const std::vector<std::shared_ptr<ParameterInfo> > & paramsInfo,
 		void* dependencyObject, const Types &... params)
 	{
 		if (cacheData.count(hash) == 0)
